add test coverage for both regex and python-exec based loading, and fix a bug encountered with the latter
