--- test/correctness/PySys_internal_171/pysystest.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<pysystest>
-	<title>Test loading - Descriptors embedded in a pysystest.py file, multiple encodings</title>
-
-	<groups inherit="true" groups=""/>
-
-	<modes inherit="true">
-	</modes>
-
-	<!-- To skip the test, uncomment this (and provide a reason): <skipped reason=""/> -->
-	<!-- To provide a bug/story/requirement id for requirements tracing, uncomment this: <requirement id=""/> -->
-</pysystest>
